Added setPressure() to simulatedReactor
The setPressure function changes the pressure to a set figure. It operates slowly to simulate a real vacuum chamber. (But I really don't know what an appropriate speed would be.) It also lowers temperature as pressure gets lowered. (Again, I don't know the actual relationship between pressure and temp. so there is a placeholder function.)